Fikset tabs til egne komponenter
